+++ src/Result.php
@@ -73,16 +73,18 @@ final class Result
      * Each value is an array of error message strings.
      *
      * @param string $separator Attribute path separator. Dot is used by default.
+     * @param string|null $escape Symbol that will be escaped with a backslash char (`\`) into path elements.
+     * When is null value path returns without escaping.
      *
      * @return array Arrays of error messages indexed by attribute path.
      *
      * @psalm-return array<string, non-empty-list<string>>
      */
-    public function getErrorMessagesIndexedByPath(string $separator = '.'): array
+    public function getErrorMessagesIndexedByPath(string $separator = '.', ?string $escape = '.'): array
     {
         $errors = [];
         foreach ($this->errors as $error) {
-            $stringValuePath = implode($separator, $error->getValuePath($separator));
+            $stringValuePath = implode($separator, $error->getValuePath($escape));
             $errors[$stringValuePath][] = $error->getMessage();
         }
 
@@ -158,12 +160,18 @@ final class Result
      *
      * @param string $attribute Attribute name.
      * @param string $separator Attribute path separator. Dot is used by default.
+     * @param string|null $escape Symbol that will be escaped with a backslash char (`\`) into path elements.
+     * When is null value path returns without escaping.
      *
      * @return array Arrays of error messages for the attribute specified indexed by attribute path.
      *
      * @psalm-return array<string, non-empty-list<string>>
      */
-    public function getAttributeErrorMessagesIndexedByPath(string $attribute, string $separator = '.'): array
+    public function getAttributeErrorMessagesIndexedByPath(
+        string $attribute,
+        string $separator = '.',
+        ?string $escape = '.',
+    ): array
     {
         $errors = [];
         foreach ($this->errors as $error) {
@@ -172,7 +180,7 @@ final class Result
                 continue;
             }
 
-            $valuePath = implode($separator, array_slice($error->getValuePath($separator), 1));
+            $valuePath = implode($separator, array_slice($error->getValuePath($escape), 1));
             $errors[$valuePath][] = $error->getMessage();
         }